Fix hollow step markers for is-small modifier
This is the SASS syntax error, which makes my SASS compiler fail to compile (Rails 5.1.3 + Webpacker 3.0). It works in other compilers, like https://www.sassmeister.com/, however the values of $steps-thin-marker-size and $steps-thin-divider-size are not evaluated, so the resulting CSS still has those variables in it, not their values.

I tested this change in my project and in https://www.sassmeister.com/ (switch to SASS syntax).